Replaced hero banner cards with MUI card component, replaced marketing image with 'horn' image, removed all page imports and converted useTheme styles to MUI shorthand
